Added postCreateUniqueRecipe(), allowing multiple runs of the app
plus code clean up.
