+++ cantabular-import/analysis/full-import-export/go.mod
@@ -3,9 +3,19 @@ module ONSdigital/full-import-export
 go 1.15
 
 require (
-	github.com/ONSdigital/dp-api-clients-go v1.37.0
-	github.com/ONSdigital/dp-kafka/v3 v3.0.0-beta.4
-	github.com/ONSdigital/log.go v1.0.1
+	github.com/ONSdigital/dp-api-clients-go v1.43.0
+	github.com/ONSdigital/dp-healthcheck v1.2.1 // indirect
+	github.com/ONSdigital/log.go v1.1.0
 	github.com/davecgh/go-spew v1.1.1
+	github.com/fatih/color v1.13.0 // indirect
+	github.com/gopherjs/gopherjs v0.0.0-20210803090616-8f023c250c89 // indirect
+	github.com/hokaccha/go-prettyjson v0.0.0-20211117102719-0474bc63780f // indirect
 	github.com/kelseyhightower/envconfig v1.4.0
+	github.com/kr/text v0.2.0 // indirect
+	github.com/mattn/go-colorable v0.1.12 // indirect
+	github.com/satori/go.uuid v1.2.0
+	github.com/smartystreets/goconvey v1.7.2 // indirect
+	golang.org/x/net v0.0.0-20211112202133-69e39bad7dc2 // indirect
+	golang.org/x/sys v0.0.0-20211124211545-fe61309f8881 // indirect
+	gopkg.in/check.v1 v1.0.0-20201130134442-10cb98267c6c // indirect
 )
